refactor: 残存するRequest構造体をhandler層からdto層に移動
learning_resource.go、answer.go、post_tag.goのRequest構造体を
dto層に統一し、handler層ではdto参照に変更。
