+++ backend/internal/dto/answer_dto.go
@@ -1,0 +1,11 @@
+package dto
+
+// CreateAnswerRequest は回答作成のリクエストボディ。
+type CreateAnswerRequest struct {
+	Body string `json:"body" binding:"required"`
+}
+
+// UpdateAnswerRequest は回答更新のリクエストボディ。
+type UpdateAnswerRequest struct {
+	Body string `json:"body" binding:"required"`
+}
