+++ backend/internal/dto/post_tag_dto.go
@@ -1,0 +1,6 @@
+package dto
+
+// SetTagsRequest はタグ設定のリクエストボディ。
+type SetTagsRequest struct {
+	Tags []string `json:"tags"`
+}
